test: cover auto-registration for every storage backend
